Split settings screen in files
